requirements.txt
PyQt6
PyQt5
PySide2
spotipy
mysql-connector
dotenv
python-dotenv
mutagen
urllib3
charset-normalizer
click
colorama
ffmpeg-python
filetype
future
mysql-connector-python
numpy
opencv-python
packaging
pygame
PyQt5-sip
pyqt5-tools
python-dotenv
qt5-applications
qt5-tools
requests
setuptools
toml
urllib3



